Fix coverity warning. (LArCalibTools-00-09-47)
	* Tagging LArCalibTools-00-09-47.
	* Fix coverity warning.

2015-01-11  scott snyder  <snyder@bnl.gov>

	* Tagging LArCalibTools-00-09-46.
	* AthenaBaseComps migration.

2014-11-21 Walter Lampl <walter.lampl@ cern.ch>
	* complete2Flat: Bugfix and better diagnostic message
	* tag LArCalibTools-00-09-45

2014-11-14 Walter Lampl <walter.lampl@ cern.ch>
	* fix include guard of LArDigits2Ntuple
	* tag LArCalibTools-00-09-44

2014-10-31 Walter Lampl <walter.lampl@ cern.ch>
	* LArHVScaleCorr2Ntuple: Use abstract interface ILArHVScaleCorr instead of LArHVScaleCorrComplete
	* LArCommConditions2Ntuple.py: Add HV Scale Correction
...
(Long ChangeLog diff - truncated)

+++ LArCalorimeter/LArCalibTools/LArCalibTools/LArCaliWaves2Ntuple.h
@@ -56,7 +56,7 @@ class LArCaliWaves2Ntuple : public  LArWaves2Ntuple
   NTuple::Item<long>    m_dac;
   NTuple::Item<long>    m_gain;
   NTuple::Item<long>    m_nPulsedCalibLines;
-  NTuple::Array<long>   m_pulsedCalibLines;
+  NTuple::Array<int>   m_pulsedCalibLines;
   NTuple::Item<double>  m_jitter;
   NTuple::Item<long>    m_corrUndo;
